Add WorkerPool.sendMessageToAllWorkers
This is an API that guarantees that a given message will be sent to *all*
workers in the pool. If the workers are busy, it will wait for the response to
their currently pending work, and send it immediately after that. If they are
not busy, it will occupy them with the given message.