workerPool.sendMessageToAllWorkers() don't need to be sequentially dependent
Previously, if you called sendMessageToAllWorkers() twice in a row, the second
send would wait for all workers to receive the message from the first.

However, this was unnecessary since we already have a queue of worker-specific
messages that the worker will shift a message off first thing when it returns
from being busy -- so waiting on other children to have received the message is
unnecessary.