Ok, no more travis support for node 0.8 *sigh*
Node 0.8 ships with a version of npm that doesn't understand the `^`
semver operator. That means travis CI (or anyone with that version of
node/npm) is incapable of `npm install`ing anything with a dependency
(or recursive dependency) that uses this semver operator.

As a result, Travis tests for 0.8 will always fail, and so they have
become exclusively noise.